+++ FINAL_FPS/Structure.pde
@@ -1,7 +1,7 @@
 public class Structure extends Obj {
   color surface;
   color outline;
-  public Structure(int x, int y, int z, int xAng, int yAng, int zAng, int[] mag, color surf, color out) {
+  public Structure(int x, int y, int z, float xAng, float yAng, float zAng, float[] mag, color surf, color out) {
     super(x, y, z, xAng, yAng, zAng, mag, 4);
     this.surface = surf;
     this.outline = out;
